Updates main nav and form-builder language switcher
