+++ public/static/locales/fr/form-builder.json
@@ -60,7 +60,7 @@
   "dropdown": "Liste déroulante",
   "duplicate": "Dupliquer",
   "edit": "Modifier",
-  "editingIn": "Editing in :",
+  "editingIn": "Editing in: [FR]",
   "editLink": "Modifier un lien  ",
   "elementActions": "Element actions",
   "email": "Courriel",
